v1.0.2: custom font for axis tick labels and bug fixes.
You can now set a label font (UIFont) for the x- and y-axis tick labels (default is UIFont.systemFont(ofSize: 8)). The fontSize property has been removed.
The default color of the line in a line chart is now primary instead of orange.
Bug fix: setting userInteraction to false in a DYLineView now works properly when the userInteraction property is set to true in the DYLineChartView modifier.